fix(dsv41): the fp8 gemv scale row is a uint8 e8m0 array, not float
The vectorized path declared the per-row scale pointer as const float*, but w_scale is the e8m0
byte array that ue8m0_to_f consumes, so the initialization failed to compile. Caught by the build
before any measurement ran, which is what the build gate is for.

+++ kernels/cuda/dsv41_kernels.cu
@@ -1245,7 +1245,7 @@ __global__ void gemm_fp8_gemv_kernel(const uint8_t* __restrict__ a,
     for (int row = blockIdx.x * nwarps + warp; row < n; row += gridDim.x * nwarps) {
         const uint8_t* wr = w + (size_t)row * k;
         const int srow = row >> 5;               // 32x32 block scale row
-        const float* wsr = w_scale + (size_t)srow * nb_k;
+        const uint8_t* wsr = w_scale + (size_t)srow * nb_k;
         float acc = 0.f;
         if (g_gemv_fp8_vec) {
             for (int g = 0; g < n_vec; ++g) {
